pyproject.toml
[project]
name = "todo-list-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Felipe Guimarães",email = "89321902+guimaraesfe1@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.13,<3.14"
dependencies = [
    "sqlalchemy (>=2.0.49,<3.0.0)",
    "fastapi[standard] (>=0.135.3,<0.136.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
    "pyjwt (>=2.12.1,<3.0.0)",
    "pydantic-settings (>=2.13.1,<3.0.0)"
]

[tool.poetry]
package-mode = false

[tool.ruff]
line-length = 80

[tool.ruff.format]
quote-style = "single"

[tool.ruff.lint]
select = ["I"]

[tool.taskipy.tasks]
dev = "uvicorn src.main:app --reload"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.15.10,<0.16.0)",
    "taskipy (>=1.14.1,<2.0.0)"
]
